Large update to dimension, maps and movement + many others
lua:
 - player movement
 - ai for npc

cpp
 - scaling of rendered images now per "level", not global
 - removed global TILESIZE - depends on image/object
 - generally clearer separation of image / physics / render dimensions
   -> every image/text has clear rect to render to
 - normalize velocity, improve pathfinding
 - removed Character class (obsolete)
 - removed blocks from MapWrapper - now blocks as separate wrapper (less
   code duplicate)
 - added Sprite and Sprite to better handle tile definition
 - various improvements to names of variables
 - added configuration centralization of animations
 - added  hitbox size to constructor of block
 - unified every float-like to double
... probably forgot some stuff

+++ Basics/button.h
@@ -15,7 +15,7 @@ public:
         UNCLICK,
         COUNT
     };
-    button(SDL_Rect pos, double ws = 2, double hs = 1);
+    button(SDL_Rect pos, int ws = 2, int hs = 1);
     button(SDL_Rect pos, SDL_Rect _clips[COUNT]);
     button(SDL_Rect pos, SDL_Rect clip);
     button(const button &b);
